fix(review): address Copilot feedback on OIDC PR
Objective:
- resolve actionable Copilot review comments on PR 146 before requesting the next automated review

Changes:
- tighten Admin modal, side-panel, SVG, and layout cleanup called out by review
- guard Passport Google Maps and Instagram OAuth configuration before external calls
- add browser-safe @cubid/auth/client contract exports and fix repository-relative README links

Validation:
- pnpm --filter @cubid/admin typecheck
- pnpm --filter @cubid/passport typecheck
- pnpm --filter @cubid/auth typecheck
- pnpm --filter @cubid/auth test
- pnpm --filter @cubid/oidc typecheck
- pnpm --filter @cubid/oidc test
- pnpm lint
- pnpm typecheck
- pnpm test
- pnpm build

Notes:
- local shell is Node 25, so pnpm emitted the existing Node 20 engine warning during validation
